Add files via upload
Jianbo's version of part 1, compare with the answer of A1Q1CMPUT229.drawio.png, I add an entity store, and change the relation between activities and products "display" to "process" and expand with several attribite such as view and check out. And i don't know whether should I write the key attibute for each entity even if it is not written.